requirements.txt
torch==2.5.1 
torchvision==0.20.1 
torchaudio==2.5.1

cachetools==5.5.2
datasets==2.20.0
einops==0.8.1
matplotlib==3.10.1
numpy==2.2.3
pandas==2.2.3
Pillow==11.1.0
pycocoevalcap==1.2
python_dateutil==2.9.0.post0
pytorch_lightning==2.3.0
PyYAML==6.0.1
radgraph==0.0.9
scikit_learn==1.6.1
scipy==1.15.2
sentence_transformers==3.0.1

torchmetrics==1.4.0.post0
tqdm==4.66.4
transformers==4.43.3
